Authenticate user against DB when in session
So far if a session exists for the user in the database, the user was
allowed to access proctected routes.
With this commit, the existence of a user will also be checked if the
user is in session, that is if the user account has been deleted and the
user was in session before, the user won't be allowed to access
protected routes anymore.